pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "session-adapters"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = []
authors = [
  { name = "Simone Tripodi", email = "simone.tripodi@terradue.com" },
  { name = "Fabrice Brito", email = "fabrice.brito@terradue.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "boto3",
  "loguru",
  "oras",
  "pydantic",
  "python-magic",
]

[project.urls]
Documentation = "https://github.com/Terradue/session-adapters#README.md"
Issues = "https://github.com/Terradue/session-adapters/issues"
Source = "https://github.com/Terradue/session-adapters"

[tool.hatch.build.targets.wheel]
packages = ["src/session_adapters"]

[tool.hatch.version]
path = "src/session_adapters/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/session_adapters tests}"

[tool.coverage.run]
source_pkgs = ["session_adapters", "tests"]
branch = true
parallel = true
omit = [
  "src/session_adapters/__about__.py",
]

[tool.coverage.paths]
session_adapters = ["src/session_adapters", "*/session-adapters/src/session_adapters"]
tests = ["tests", "*/session-adapters/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]


[tool.hatch.envs.test]
description = "Test environment"
dependencies = [
  "pytest>=8.0",
]

[tool.hatch.envs.test.scripts]
test = "pytest"
test-q = "pytest -q"
test-cov = "pytest --cov=opensearch_gateway"
test-show-setup = "pytest --setup-show"

[tool.hatch.envs.dev]
description = "Development environment"
dependencies = [
  "pytest>=8.0",
  "ruff",
]


[tool.hatch.envs.dev.scripts]
lint = "ruff format ."
check = "ruff check --fix"
